Layout should be on [locale] level
