+++ apps/researcher/src/app/[locale]/navigation.tsx
